+++ gz/official_replace_config.php
@@ -5,7 +5,7 @@
  */
 
 return [
-    'version' => '4.0.3',
+    'version' => '4.0.4',
     'update_date' => '2026-07-15',
     'enabled' => true,
     'default_site' => '量子',